# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Notification": {
        "description": "Something that happened, addressed to one account.\n\nAttributes:\n    id (Optional[str]): Identifier, populated on read from the store.\n    recipient_id (str): The account it is addressed to.\n    kind (NotificationKind): What it is about.\n    title (str): The one-line summary a list shows.\n    body (Optional[str]): The detail, shown when the row is opened.\n    quote_id (Optional[str]): The quote it points at, when it points at one.\n    is_read (bool): Whether the recipient has seen it.\n    created_at (Optional[datetime]): When the event happened.\n    read_at (Optional[datetime]): When the recipient marked it read.\n\nNotes:\n    - **One row per recipient, not one row per event.** A quote submitted to\n      an agency with three managers produces three notifications. Storing\n      one event and computing recipients on read would make \"have I read\n      this?\" unanswerable \u2014 read state belongs to a person, and two managers\n      must be able to disagree about whether they have dealt with something.\n    - The notification is a **record**, not a delivery mechanism. It is\n      written to the database first and pushed over the wire second, so a\n      reader who was offline, or whose stream dropped, still finds it\n      waiting. A design that only pushed would lose the notification exactly\n      when it matters most \u2014 overnight, when nobody is watching.\n    - The text is stored, not templated at read time. What a manager was\n      told is what they were told; regenerating it later would rewrite\n      history after a customer is renamed or a quote is repriced.",
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The detail, shown when the row is opened.",
            "title": "Body"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the event happened.",
            "title": "Created At"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identifier, populated on read from the store.",
            "title": "Id"
          },
          "is_read": {
            "default": false,
            "description": "Whether the recipient has seen it.",
            "title": "Is Read",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/NotificationKind",
            "description": "What it is about."
          },
          "quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The quote it points at, when it points at one.",
            "title": "Quote Id"
          },
          "read_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the recipient marked it read.",
            "title": "Read At"
          },
          "recipient_id": {
            "description": "The account it is addressed to.",
            "title": "Recipient Id",
            "type": "string"
          },
          "title": {
            "description": "The one-line summary a list shows.",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "recipient_id",
          "kind",
          "title"
        ],
        "title": "Notification",
        "type": "object"
      },
      "NotificationKind": {
        "description": "Enumeration for what a notification is about.\n\nAttributes:\n    QUOTE_SUBMITTED (str): An assistant sent a quote for validation.\n    QUOTE_VALIDATED (str): A manager approved a submitted quote.\n    QUOTE_REFUSED (str): A manager sent a submitted quote back.\n    PLANNING_COMPLETED (str): A planning run finished and calendars moved.\n\nNotes:\n    The kind is what a client keys its icon, colour and deep link off, so it\n    is an enumeration rather than free text. A notification whose kind\n    nothing recognises is one the interface cannot route the reader\n    anywhere useful from.",
        "enum": [
          "quote-submitted",
          "quote-validated",
          "quote-refused",
          "planning-completed"
        ],
        "title": "NotificationKind",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Quoting and intervention planning for a home-care agency: customers, assistants, intervention types, quotes and the planning computation.",
    "title": "rt-erp API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "description": "Return the caller's own notifications, newest first.\n\nArgs:\n    page (int): One-based page number.\n    size (int): Page size.\n    unread_only (bool): Restrict to notifications not yet read.\n    service (NotificationService): The notification service.\n    caller (User): The authenticated caller.\n\nReturns:\n    List[Notification]: The caller's notifications.\n\nNotes:\n    There is no way to read anybody else's queue, and no parameter that\n    names a recipient. The recipient is the credential.",
        "operationId": "list_notifications_api_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unread_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Unread Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  },
                  "title": "Response List Notifications Api V1 Notifications Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/v1/notifications/stream": {
      "get": {
        "description": "Stream the caller's notifications as Server-Sent Events.\n\nArgs:\n    request (Request): The incoming request, watched for disconnection.\n    token (str): A short-lived stream token from\n        ``POST /api/v1/auth/stream-token``.\n    auth (AuthService): The authentication service.\n    broadcaster (NotificationBroadcaster): The in-process fan-out.\n\nReturns:\n    StreamingResponse: An ``text/event-stream`` response.\n\nRaises:\n    MTAuthInvalidToken: If the token is missing, expired, or not scoped for\n        a stream; answered as a 401.\n\nNotes:\n    - **The credential is in the query string because it has to be.**\n      ``EventSource`` cannot set an ``Authorization`` header. What travels\n      there is not the session token but a token that lives for a minute\n      and is refused everywhere else, so a URL captured in a proxy log is\n      worth nothing by the time anybody reads it.\n    - This route authenticates itself rather than relying on the\n      middleware, and is exempt from it \u2014 the middleware only understands\n      bearer headers, and would leave ``request.state.user`` unset here.\n    - A keep-alive comment goes out every ``KEEPALIVE_SECONDS`` so that an\n      idle stream is not closed by a proxy. A comment frame is ignored by\n      ``EventSource`` and costs three bytes.",
        "operationId": "stream_notifications_api_v1_notifications_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/v1/notifications/{notification_id}/read": {
      "post": {
        "description": "Mark one of the caller's notifications as read.\n\nArgs:\n    notification_id (str): The notification to mark.\n    service (NotificationService): The notification service.\n    caller (User): The authenticated caller.\n\nReturns:\n    Notification: The updated notification.\n\nRaises:\n    MTNotificationNotFound: If it does not exist or is addressed to\n        somebody else; answered as a 404 either way.",
        "operationId": "mark_notification_read_api_v1_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Notification"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Notification Read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  }
}